+++ backend/blog-backend/src/main/java/com/kisesaki/blog/auth/service/AuthService.java
@@ -1,4 +1,4 @@
-package com.kisesaki.blog.auth;
+package com.kisesaki.blog.auth.service;
 
 import java.time.LocalDateTime;
 import java.util.Map;
